+++ changelog.md
@@ -1,6 +1,18 @@
 # Jam Changelog
 
-## Version 1.0.0
+## Version 1.0.3 (2016-12-22)
+
+-   Make repo in front end controller protected rather than private
+
+## Version 1.0.2 (2016-12-21)
+
+-   Fix nested entity list redirect 
+
+## Version 1.0.1 (2016-12-16)
+
+-   Fix entity list save redirect for nested lists
+
+## Version 1.0.0 (2016-12-16)
 
 -   Add max_uses to templates
 -   Don't make slugs unique